requirements.txt
async-timeout==4.0.2
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Deprecated==1.2.13
filelock==3.12.0
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Session==0.4.0
fsspec==2023.4.0
gcloud==0.18.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
huggingface-hub==0.14.1
idna==2.7
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.2
numpy==1.24.3
oauth2client==4.1.3
packaging==23.1
protobuf==4.23.0rc2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dotenv==1.0.0
python-jwt==4.0.0
PyYAML==6.0
redis==4.5.4
regex==2023.3.23
requests==2.19.1
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
typing_extensions==4.5.0
urllib3==1.23
Werkzeug==2.3.2
wrapt==1.15.0
